Demo compiler error with incorrect file path
This project demonstrates a bug in the Elm Language Server with projects
that have "source-directories" entries that start with "..". When a source
file in such a path contains a compiler error, the Elm Make Diagnostics
in the Elm Language Server handle the file path incorrectly.

Specifically, [these lines][src] interpret the leading "." on the file
path as a "current directory" (.), and strip that leading "." off the
path, converting "../lib/Library.elm" to "./lib/Library.elm" (which is
incorrect).

[src]: https://github.com/elm-tooling/elm-language-server/blob/39f399d0bbd50a04cd966f32f73fa1e1bdf96813/src/providers/diagnostics/elmDiagnosticsHelper.ts#L38-L42